requirements.txt
flask>=2.3.0
google-generativeai>=0.8.5
pillow>=9.5.0
python-dotenv>=1.0.0
werkzeug>=2.3.0 